Fix compilation warnings under win32
Andi, it can be rounding problem here between double => int cast

svn path=/trunk/extragear/graphics/kipi-plugins/; revision=883914

+++ removeredeyes/blobsettingsbox.cpp
@@ -91,12 +91,12 @@ void BlobSettingsBox::setMinBlobSize(int value)
 
 double BlobSettingsBox::minRoundness() const
 {
-    return d->minRoundnessNumInput->value();
+    return ((double)(d->minRoundnessNumInput->value()));
 }
 
 void BlobSettingsBox::setMinRoundness(double value)
 {
-    d->minRoundnessNumInput->setValue(value);
+    d->minRoundnessNumInput->setValue((int)value);
 }
 
 } // namespace KIPIRemoveRedEyesPlugin
